login control - 20170813

+++ src/main/java/com/stemcloud/smart/web/dao/SensorInfoRepository.java
@@ -16,7 +16,7 @@ import java.util.List;
 public interface SensorInfoRepository extends CrudRepository<SensorInfo, Integer> {
     List<SensorInfo> findByIsShare(int isShare);
 
-    List<SensorInfo> findByAppId(int appId);
+    List<SensorInfo> findByAppId(long appId);
 
     @Query(value = "UPDATE base_sensor_info SET sensor_name = :name, " +
             "sensor_code = :code, sensor_type = :sType, longitude = :longitude, latitude = :latitude, sensor_city = :city, " +
